회원 클래스 age int으로 변경

+++ src/main/java/oss/servlet/domain/member/Member.java
@@ -8,12 +8,12 @@ public class Member {
 
     private Long id ; // 회원아이디
     private String username; //회원 이름
-    private Long age ; // 나이
+    private int age ; // 나이
 
     public Member() { // 기본생성자
     }
 
-    public Member(String username, Long age) {
+    public Member(String username, int age) {
         this.username = username;
         this.age = age;
     }
